allow user to remove poll
