+++ feature/post/src/main/java/org/mozilla/social/post/bottombar/BottomBar.kt
@@ -227,12 +227,3 @@ private fun CharacterCountLabel(characterCountText: String) {
         color = MoSoTheme.colors.textSecondary,
     )
 }
-
-data class BottomBarState(
-    val imageButtonEnabled: Boolean = false,
-    val videoButtonEnabled: Boolean = false,
-    val pollButtonEnabled: Boolean = false,
-    val contentWarningText: String? = null,
-    val characterCountText: String = "",
-    val maxImages: Int = NewPostViewModel.MAX_IMAGES,
-)
